[Refactor] 외부 API 연동 코드를 infrastructure 패키지로 구조 개선
