+++ src/main/java/verbly/spring/domain/payment/service/SubscriptionScheduler.java
@@ -8,7 +8,7 @@ import org.springframework.transaction.annotation.Transactional;
 import verbly.spring.domain.payment.entity.Subscription;
 import verbly.spring.domain.payment.enums.SubscriptionStatus;
 import verbly.spring.domain.payment.repository.SubscriptionRepository;
-import verbly.spring.infrastructure.kakao.KakaoPayClient;
+import verbly.spring.global.infrastructure.kakao.KakaoPayClient;
 
 import java.time.LocalDateTime;
 import java.util.List;
