+++ src/main/java/verbly/spring/domain/payment/service/SubscriptionService.java
@@ -13,8 +13,8 @@ import verbly.spring.domain.payment.repository.SubscriptionRepository;
 import verbly.spring.domain.user.entity.User;
 import verbly.spring.domain.user.repository.UserRepository;
 import verbly.spring.global.common.code.ErrorStatus;
-import verbly.spring.infrastructure.kakao.KakaoPayClient;
-import verbly.spring.infrastructure.kakao.dto.KakaoPayDTO;
+import verbly.spring.global.infrastructure.kakao.KakaoPayClient;
+import verbly.spring.global.infrastructure.kakao.dto.KakaoPayDTO;
 
 import java.time.LocalDateTime;
