+++ src/main/java/verbly/spring/domain/user/service/UserCommandServiceImpl.java
@@ -16,7 +16,7 @@ import verbly.spring.domain.user.repository.ProfileImageRepository;
 import verbly.spring.domain.user.repository.UserRepository;
 import verbly.spring.domain.uuid.entity.Uuid;
 import verbly.spring.domain.uuid.repository.UuidRepository;
-import verbly.spring.global.common.aws.s3.AmazonS3Manager;
+import verbly.spring.global.infrastructure.aws.s3.AmazonS3Manager;
 import verbly.spring.global.common.code.ErrorStatus;
 
 import java.util.UUID;
