--- src/main/java/verbly/spring/global/common/dto/openAI/OpenAIMessage.java
@@ -1,17 +1,0 @@
-package verbly.spring.global.common.dto.openAI;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class OpenAIMessage {
-    private String role;    // system, user, assistant
-    private String content; // message content
-
-    public OpenAIMessage(String role, String content) {
-        this.role = role;
-        this.content = content;
-    }
-}
