+++ src/main/java/verbly/spring/global/infrastructure/openAI/dto/OpenAIResponseDTO.java
@@ -1,0 +1,19 @@
+package verbly.spring.global.infrastructure.openAI.dto;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class OpenAIResponseDTO {
+    private List<Choice> choices;
+
+    @Getter
+    @NoArgsConstructor(access = AccessLevel.PROTECTED)
+    public static class Choice {
+        private OpenAIMessage message;
+    }
+}
